separates svo downloader to provider
